Rename Benchmark Now it reflects name of the tested RandomExecutor Also moved SyncRuntime into it
